Renamed the test folders with the environment type logic instead.
