+++ mipengine/algorithms/logistic_regression.py
@@ -265,7 +265,7 @@ def sum_tensors(xs):
 
 
 def test_logistic_regression():
-    data = pd.read_csv("tests/integration_tests/data/dementia/edsd.csv")
+    data = pd.read_csv("tests/dev_env_tests/data/dementia/edsd.csv")
     y_name = "alzheimerbroadcategory"
     x_names = [
         "lefthippocampus",
