+++ tests/dev_env_tests/deployment_template.toml
@@ -6,7 +6,7 @@ rabbitmq_image = "madgik/mipengine_rabbitmq:latest"
 
 cdes_metadata_path = "./tests/demo_data"
 
-algorithm_folders = "./tests/algorithms"
+algorithm_folders = ["./mipengine/algorithms,./tests/algorithms"]
 
 node_registry_update_interval = 20
 celery_tasks_timeout = 10
