+++ tests/dev_env_tests/test_tables.py
@@ -5,8 +5,8 @@ from mipengine.node_tasks_DTOs import ColumnInfo
 from mipengine.datatypes import DType
 from mipengine.node_tasks_DTOs import TableData
 from mipengine.node_tasks_DTOs import TableSchema
-from tests.integration_tests.nodes_communication import get_celery_app
-from tests.integration_tests.nodes_communication import get_celery_task_signature
+from tests.dev_env_tests.nodes_communication import get_celery_app
+from tests.dev_env_tests.nodes_communication import get_celery_task_signature
 
 local_node = get_celery_app("localnode1")
 local_node_create_table = get_celery_task_signature(local_node, "create_table")
