+++ tests/dev_env_tests/test_views_and_filters.py
@@ -7,8 +7,8 @@ from mipengine.datatypes import DType
 from mipengine.node_tasks_DTOs import InsufficientDataError
 from mipengine.node_tasks_DTOs import TableData
 from mipengine.node_tasks_DTOs import TableSchema
-from tests.integration_tests.nodes_communication import get_celery_task_signature
-from tests.integration_tests.nodes_communication import get_celery_app
+from tests.dev_env_tests.nodes_communication import get_celery_task_signature
+from tests.dev_env_tests.nodes_communication import get_celery_app
 
 local_node_id = "localnode1"
 local_node = get_celery_app(local_node_id)
