+++ tests/prod_env_tests/kind_configuration/kind_cluster.yaml
@@ -6,7 +6,7 @@ nodes:
   - containerPort: 30000
     hostPort: 5000
   extraMounts:
-  - hostPath: ./tests/e2e_tests/kind_configuration/hostname_master
+  - hostPath: ./tests/prod_env_tests/kind_configuration/hostname_master
     containerPath: /etc/hostname
   - hostPath: ./tests/demo_data
     containerPath: /opt/data
@@ -14,13 +14,13 @@ nodes:
     containerPath: /opt/monetdb
 - role: worker
   extraMounts:
-  - hostPath: ./tests/e2e_tests/kind_configuration/hostname_worker1
+  - hostPath: ./tests/prod_env_tests/kind_configuration/hostname_worker1
     containerPath: /etc/hostname
   - hostPath: /opt/monetdb1
     containerPath: /opt/monetdb
 - role: worker
   extraMounts:
-  - hostPath: ./tests/e2e_tests/kind_configuration/hostname_worker2
+  - hostPath: ./tests/prod_env_tests/kind_configuration/hostname_worker2
     containerPath: /etc/hostname
   - hostPath: /opt/monetdb2
     containerPath: /opt/monetdb
