+++ tests/prod_env_tests/test_post_algorithms.py
@@ -4,7 +4,7 @@ import re
 import pytest
 import requests
 
-from tests.e2e_tests import algorithms_url
+from tests.prod_env_tests import algorithms_url
 
 
 def get_parametrization_list_success_cases():
