+++ tests/standalone_tests/test_algorithm_result.py
