+++ tests/standalone_tests/test_algorithms_folder.py
