+++ tests/standalone_tests/test_node_registry.py
