+++ tox.ini
@@ -7,4 +7,4 @@ isolated_build = True
 whitelist_externals = poetry
 commands =
     poetry install -v
-    poetry run pytest -n5 tests/unit_tests
+    poetry run pytest -n5 tests/standalone_tests
